move firebase to the file that use it
